Finalizando o gerador aleatório de pessoas

+++ Model/ViewModel/ClienteViewModel.cs
@@ -4,7 +4,7 @@
     {
         public string Nome { get; set; }
         public string DataNascimento { get; set; }
-        public short AgenciaBancaria { get; set; }
+        public int AgenciaBancaria { get; set; }
         public int ContaBancaria { get; set; }
     }
 }
